pyproject.toml
[project]
name = "rspcp-bachelors-thesis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.18.4",
    "duckdb>=1.4.4",
    "duckdb-engine>=0.17.0",
    "isort>=8.0.1",
    "loguru>=0.7.3",
    "mkdocs>=1.6.1",
    "mkdocs-api-autonav>=0.4.0",
    "mkdocs-material>=9.7.1",
    "mkdocstrings>=1.0.3",
    "mlflow>=3.9.0",
    "pre-commit>=4.5.1",
    "pydantic-settings>=2.13.1",
    "python-binance>=1.0.29",
    "sqlalchemy>=2.0.46",
    "tenacity>=9.1.2",
    "typer>=0.24.1",
]

[tool.pyright]
include = ["src"]
exclude = ["src/tests/*"]
venvPath = "."
venv = ".venv"
reportMissingImports = true
reportMissingTypeStubs = true
reportPrivateUsage = true

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv"
]

# Path to store cache
cache-dir = "~/.cache/ruff"

# Fix by default when run ruff
fix = true

# A list of file patterns to include when linting
include = ["*.py"]

# Style in which violation messages should be formatted
output-format = "concise"

# Hints how to fix lint violations
show-fixes = true

# Same as Black.
line-length = 119
indent-width = 4



[tool.ruff.lint]
# Ruff will raise warning f.e. if you use cyrylica "a" instead of unicode
allowed-confusables = []

# A list of rule codes or prefixes to consider fixable
fixable = ["ALL"]

# A list of rule codes or prefixes to ignore
ignore = [
    "I", "ISC", "RSE102", "EM101", "TRY003", "EM102"
]

# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`)  codes by default.
# Unlike Flake8, Ruff doesn't enable pycodestyle warnings (`W`) or
# McCabe complexity (`C901`) by default.
select = [
    "F", "FAST", "ANN", "ASYNC", "C4", "DTZ", "EM", "INP", "PIE", "T20", "PT", "Q", "RSE", "RET",
    "SIM", "TID", "ARG", "PTH", "N", "PERF", "DOC", "D", "PLR", "UP", "TRY", "S", "B", "E"
]

# Comments starting with these tags will be ignored
task-tags = ["TODO"]

# A list of rule codes or prefixes to consider non-fixable
unfixable = []

extend-select = ["S", "B", "E", "PLR", "UP"]
preview = true

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"src/tests/*" = [
    "S101",    # assert is fine in tests
    "S105",    # hardcoded fake passwords are intentional
    "S106",    # hardcoded fake passwords are intentional
    "PLR6301", # test methods don't use self — pytest requires instance methods
    "PLR2004", # magic values in assertions are idiomatic (assert len(x) == 2)
    "ARG002",  # unused fixture args trigger side-effects (e.g. set_binance_env)
    "DOC501",  # test docstrings don't need Raises sections
    "PT012",   # multi-statement pytest.raises blocks are acceptable
    "PT017",   # assertion on exception in except block (testing chaining)
    "TRY301",  # raise-in-try is the point of exception tests
]
"**/alembic/versions/*.py" = [
    "D103"
]

[tool.ruff.format]
# Enable reformatting of code snippets in docstrings.
docstring-code-format = true

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Prefer double quotes over double quotes.
quote-style = "double"

[tool.isort]
# Skip files and directories
skip = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
    "tests"
]

# Line length for imports
line_length = 119

# Imports order by section name
sections = ['FUTURE', 'STDLIB', 'THIRDPARTY', 'FIRSTPARTY', 'LOCALFOLDER']

# Groupping style
multi_line_output = 5

# Sort every section in alphabetical order
force_alphabetical_sort_within_sections = true

# Sort from imports in alphabetical order
no_inline_sort = false

# Lines after imports
lines_after_imports = 2

[tool.pytest.ini_options]
filterwarnings = [
    "ignore::DeprecationWarning:websockets",
    "ignore::DeprecationWarning:binance",
]

[dependency-groups]
dev = [
    "isort>=8.0.1",
    "pyright>=1.1.408",
    "pytest>=9.0.2",
    "ruff>=0.15.4",
]
